refactor: 💡 Move constants to a separate file
